cast to int, as Bret recommends

+++ source/event_loop.c
@@ -132,7 +132,7 @@ static struct aws_event_loop_group *s_event_loop_group_new(
 
             /* Thread name should be <= 15 characters */
             char thread_name[32] = {0};
-            int thread_name_len = snprintf(thread_name, sizeof(thread_name), "AwsEventLoop %d", i + 1);
+            int thread_name_len = snprintf(thread_name, sizeof(thread_name), "AwsEventLoop %d", (int)i + 1);
             if (thread_name_len > AWS_THREAD_NAME_RECOMMENDED_STRLEN) {
                 snprintf(thread_name, sizeof(thread_name), "AwsEventLoop");
             }
